requirements.txt
jsonschema
google-generativeai
